Improved struct field names.
The name changes are a bit naughty in that they no longer match
exactly the names from the documentation. I found the docs to be
ambiguous (and sometimes misleading), so opted to use the names that
make sense to me. If this library gains popularity for some reason then
I will revisit this decision.